Broken commit: Attempt made at firebase auth still some problems with it though
