+++ package.json
@@ -7,13 +7,18 @@
     "@material-ui/core": "^4.5.2",
     "@material-ui/icons": "^4.5.1",
     "@material-ui/pickers": "^3.2.7",
+    "firebase": "^7.5.0",
     "moment": "^2.24.0",
     "react": "^16.11.0",
     "react-beautiful-dnd": "^12.0.0",
     "react-dom": "^16.11.0",
+    "react-firebaseui": "^4.0.0",
+    "react-redux": "^7.1.3",
+    "react-redux-firebase": "^3.0.5",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
+    "redux": "^4.0.4",
     "styled-components": "^4.4.1"
   },
   "scripts": {
