+++ src/auth/reducers.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import { firebaseReducer } from 'react-redux-firebase';
+
+const rootReducer = combineReducers({
+  firebase: firebaseReducer,
+});
+
+export default rootReducer;
