+++ src/auth/store.js
@@ -1,0 +1,8 @@
+import { createStore } from 'redux';
+import rootReducer from './reducers';
+
+// Create store with reducers and initial state
+const initialState = {};
+const store = createStore(rootReducer, initialState);
+
+export default store;
